+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Load the twomemo backend even though SCE is not supported
 - Treat malformed device list as empty instead of an unrecoverable error
+- Try uploading the twomemo bundle again without MAX_ITEMS if that option is not supported
 
 ## [1.1.0] - 7th of October, 2024